requirements.txt
# ================= WEB FRAMEWORK =================
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0

# ================= DATABASE =================
sqlalchemy>=2.0.30,<2.1.0
asyncpg>=0.29.0,<0.30.0
alembic>=1.13.0,<1.15.0

# ================= VALIDATION & PYDANTIC =================
# Ключевое: разрешаем более новые версии pydantic-core с колесами под 3.14
pydantic>=2.9.0,<2.11.0
pydantic-core>=2.23.0,<2.28.0
pydantic[email]>=2.9.0,<2.11.0
email-validator>=2.2.0,<2.3.0

# ================= SECURITY & AUTH =================
passlib[bcrypt]>=1.7.4,<1.8.0
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]>=3.3.0,<3.4.0
python-multipart>=0.0.12,<0.1.0

# ================= ENVIRONMENT & CONFIG =================
python-dotenv>=1.0.0,<1.1.0

# ================= TEMPLATES & FORMS =================
jinja2>=3.1.0,<3.2.0
wtforms>=3.1.0,<3.2.0

# ================= ADMIN PANEL =================
# sqladmin 0.19+ совместим с SQLAlchemy 2.0
sqladmin[full]>=0.19.0,<0.20.0

# ================= IMAGE PROCESSING =================
# Pillow 10.4+ имеет колеса под новые версии Python
pillow>=10.4.0,<11.0.0

# ================= DEVELOPMENT & TESTING =================
ruff>=0.7.0,<0.9.0
pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<0.25.0
httpx>=0.27.0,<0.28.0

# ================= UTILITIES =================
tenacity>=9.0.0,<10.0.0